ap-5505: refactor change_of_names controller, route, view and flow step
Move out of the proceedings_sca namespace so that it can be shared by non-sca proceedings
