+++ app/services/flow/steps/proceedings_sca/heard_togethers_step.rb
@@ -8,7 +8,7 @@ module Flow
 
           case options[:proceeding].ccms_code
           when "PB019", "PB020", "PB023", "PB024"
-            :proceedings_sca_change_of_names
+            :change_of_names
           else
             :has_other_proceedings
           end
