+++ app/services/flow/steps/provider_start/change_of_names_step.rb
@@ -1,6 +1,6 @@
 module Flow
   module Steps
-    module ProceedingsSCA
+    module ProviderStart
       ChangeOfNamesStep = Step.new(
         path: ->(application) { Steps.urls.providers_legal_aid_application_change_of_names_path(application) },
         forward: :has_other_proceedings,
