+++ config/locales/en/providers.yml
@@ -1813,10 +1813,6 @@ en:
           page_title:
             single: Will this proceeding be heard as an alternative to the '%{proceeding}' proceeding?
             multiple: Will this proceeding be heard as an alternative to any of the special children act proceedings?
-      change_of_names:
-        error: Select yes if this proceeding is for a change of name application
-        show:
-          heading: Is this proceeding for a change of name application?
       interrupts:
         show:
           proceeding_issue_status: &default
@@ -1864,6 +1860,10 @@ en:
           yes_button: Yes, delete it
           no_button: No, do not delete it
           warning: You cannot undo this action
+    change_of_names:
+        error: Select yes if this proceeding is for a change of name application
+        show:
+          heading: Is this proceeding for a change of name application?
     proceeding_loop:
       single_proceeding: Proceeding 1
       multi_proceeding: Proceeding %{position} of %{total}
