+++ spec/requests/providers/change_of_names_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Providers::ProceedingsSCA::ChangeOfNamesController do
+RSpec.describe Providers::ChangeOfNamesController do
   let(:legal_aid_application) { create(:legal_aid_application, :with_proceedings, explicit_proceedings: %i[da001]) }
   let(:login_provider) { login_as legal_aid_application.provider }
 
@@ -13,7 +13,7 @@ RSpec.describe Providers::ProceedingsSCA::ChangeOfNamesController do
     end
 
     it "renders show with ok status" do
-      expect(response).to have_http_status(:ok).and render_template("providers/proceedings_sca/change_of_names/show")
+      expect(response).to have_http_status(:ok).and render_template("providers/change_of_names/show")
     end
 
     context "when not authenticated" do
